Task6 Built service to download coins

+++ InvestCrypto/InvestCrypto/Models/CoinModel.swift
@@ -71,7 +71,7 @@ import Foundation
 
 struct CoinModel: Identifiable, Codable {
     let id: String
-        let symbol: String
+    let symbol: String
     let name: String
     let image: String
     let currentPrice: Double
@@ -80,7 +80,7 @@ struct CoinModel: Identifiable, Codable {
     let priceChange24H, priceChangePercentage24H: Double?
     let marketCapChange24H: Double?
     let marketCapChangePercentage24H: Double?
-    let circulatingSupply, totalSupply, maxSupply, ath: Int?
+    let circulatingSupply, totalSupply, maxSupply, ath: Double?
     let athChangePercentage: Double?
     let athDate: String?
     let atl, atlChangePercentage: Double?
